# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Knowledge Inbox Server",
    "version": "0.2.0"
  },
  "paths": {
...
    "/v1/asr-settings": {
      "get": {
        "tags": [
          "asr"
        ],
        "summary": "Get Asr Settings",
        "operationId": "get_asr_settings_v1_asr_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "asr"
        ],
        "summary": "Put Asr Settings",
        "operationId": "put_asr_settings_v1_asr_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AsrSettingsIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/items/{item_id}/asr": {
      "post": {
        "tags": [
          "asr"
        ],
        "summary": "Trigger Asr",
        "description": "手动触发音频转写；模型别名/选中的音频幂等，重复点击不重复建任务。",
        "operationId": "trigger_asr_v1_items__item_id__asr_post",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/AsrTriggerIn"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "asr"
        ],
        "summary": "Get Asr Status",
        "operationId": "get_asr_status_v1_items__item_id__asr_get",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/items/{item_id}/audio-sources": {
      "get": {
        "tags": [
          "asr"
        ],
        "summary": "Get Audio Sources",
        "description": "页面音频候选：Worker 首次解析后缓存；返回 pending/ready/unsupported。\n\n只有一个逻辑音频时无需选择（ready 且 candidates 为空即\"可直接转写\"）。",
        "operationId": "get_audio_sources_v1_items__item_id__audio_sources_get",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Audio Sources V1 Items  Item Id  Audio Sources Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/items/{item_id}/asr/cancel": {
      "post": {
        "tags": [
          "asr"
        ],
        "summary": "Cancel Asr",
        "operationId": "cancel_asr_v1_items__item_id__asr_cancel_post",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AsrSettingsIn": {
        "properties": {
          "auto_when_no_track": {
            "type": "boolean",
            "title": "Auto When No Track"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "auto_when_no_track"
        ],
        "title": "AsrSettingsIn"
      },
      "AsrTriggerIn": {
        "properties": {
          "retry_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry Model"
          },
          "audio_candidate_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio Candidate Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "AsrTriggerIn"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
  "servers": [
    {
      "url": "https://kb.jerrythealpaca.cn"
    }
  ]
}